switched a lot of uint8_t variables to int to avoid extra masking instructions when working with non natural types on 32bit platform;

+++ Inc/MIDI_application.h
@@ -35,8 +35,8 @@ MIDI_ApplicationTypeDef;
 extern USBH_HandleTypeDef hUsbHostFS;
 extern MIDI_ApplicationTypeDef Appli_state;
 extern uint8_t MIDI_RX_Buffer[2][RX_BUFF_SIZE] __ATTR_RAM_D2; // MIDI reception buffer
-extern uint8_t MIDI_read_buffer;
-extern uint8_t MIDI_write_buffer;
+extern int MIDI_read_buffer;
+extern int MIDI_write_buffer;
 /* Exported functions ------------------------------------------------------- */
 
 void MIDI_Application(void);
